neuroquantum_core::quantum_processor

Trait Oracle

pub trait Oracle: Send + Sync {
    // Required methods
    fn is_target(&self, index: usize) -> bool;
    fn apply_phase_flip(&self, state_vector: &mut [Complex64]);
}

Oracle trait for Grover’s algorithm - marks target states

Required Methods

fn is_target(&self, index: usize) -> bool

Check if the given index matches the search criterion

fn apply_phase_flip(&self, state_vector: &mut [Complex64])

Apply phase flip to target states in the quantum state vector